# [show-gpus] Optimize show-gpus for GCP and reorganize TPU