… sky/clouds/service_catalog/ibm_catalog.py
… sky/clouds/service_catalog/kubernetes_catalog.py
… sky/clouds/service_catalog/lambda_catalog.py
… sky/clouds/service_catalog/oci_catalog.py
… sky/clouds/service_catalog/runpod_catalog.py
… sky/clouds/service_catalog/scp_catalog.py
… sky/clouds/service_catalog/vsphere_catalog.py
… tests/test_list_accelerators.py