<!-- Describe the changes in this PR -->

Fixes #2953.

This is to reimplement the fix in #3083 reverted in #3110. This PR makes the implementation have similar speed as before with correct results.

This PR:
```
time sky show-gpus
1.470s

time sky show-gpus --cloud gcp
0.991s

time sky show-gpus v100
1.431s

time sky show-gpus v100 --cloud gcp
1.041s

time sky show-gpus a100
1.629s
```
```
sky show-gpus tpu
GPU         QTY  CLOUD  INSTANCE_TYPE  DEVICE_MEM  vCPUs  HOST_MEM  HOURLY_PRICE  HOURLY_SPOT_PRICE  REGION        
tpu-v2-8    1    GCP    (attachable)   -           -      -         $ 4.500       $ 1.350            us-central1   
tpu-v2-32   1    GCP    (attachable)   -           -      -         $ 24.000      $ 7.200            europe-west4  
tpu-v2-128  1    GCP    (attachable)   -           -      -         $ 96.000      $ 28.800           europe-west4  
tpu-v2-256  1    GCP    (attachable)   -           -      -         $ 192.000     $ 57.600           europe-west4  
tpu-v2-512  1    GCP    (attachable)   -           -      -         $ 384.000     $ 115.200          europe-west4  
...
```

Master 905dd9c
```
time sky show-gpus
1.833s

time sky show-gpus --cloud gcp
1.332s

time sky show-gpus v100
1.368s

time sky show-gpus v100 --cloud gcp
0.975s

time sky show-gpus a100
1.430s
```

```
 time sky show-gpus tpu
GPU         QTY  CLOUD  INSTANCE_TYPE  DEVICE_MEM  vCPUs  HOST_MEM  HOURLY_PRICE  HOURLY_SPOT_PRICE  REGION        
tpu-v2-128  1    GCP    TPU-VM         -           -      -         $ 96.000      $ 28.800           europe-west4  

GPU         QTY  CLOUD  INSTANCE_TYPE  DEVICE_MEM  vCPUs  HOST_MEM  HOURLY_PRICE  HOURLY_SPOT_PRICE  REGION        
tpu-v2-256  1    GCP    TPU-VM         -           -      -         $ 192.000     $ 57.600           europe-west4  

GPU        QTY  CLOUD  INSTANCE_TYPE  DEVICE_MEM  vCPUs  HOST_MEM  HOURLY_PRICE  HOURLY_SPOT_PRICE  REGION        
tpu-v2-32  1    GCP    TPU-VM         -           -      -         $ 24.000      $ 7.200            europe-west4  

GPU         QTY  CLOUD  INSTANCE_TYPE  DEVICE_MEM  vCPUs  HOST_MEM  HOURLY_PRICE  HOURLY_SPOT_PRICE  REGION        
tpu-v2-512  1    GCP    TPU-VM         -           -      -         $ 384.000     $ 115.200          europe-west4  

GPU       QTY  CLOUD  INSTANCE_TYPE  DEVICE_MEM  vCPUs  HOST_MEM  HOURLY_PRICE  HOURLY_SPOT_PRICE  REGION       
tpu-v2-8  1    GCP    TPU-VM         -           -      -         $ 4.500       $ 1.350            us-central1  

GPU          QTY  CLOUD  INSTANCE_TYPE  DEVICE_MEM  vCPUs  HOST_MEM  HOURLY_PRICE  HOURLY_SPOT_PRICE  REGION        
tpu-v3-1024  1    GCP    TPU-VM         -           -      -         $ 1024.000    $ 307.200          europe-west4  

GPU         QTY  CLOUD  INSTANCE_TYPE  DEVICE_MEM  vCPUs  HOST_MEM  HOURLY_PRICE  HOURLY_SPOT_PRICE  REGION        
tpu-v3-128  1    GCP    TPU-VM         -           -      -         $ 128.000     $ 38.400           europe-west4 
...
```

<!-- Describe the tests ran -->
<!-- Unit tests (tests/test_*.py) are part of GitHub CI; below are tests that launch on the cloud. -->

Tested (run the relevant ones):

- [ ] Code formatting: `bash format.sh`
- [ ] Any manual or new tests for this PR (please specify below)
- [ ] All smoke tests: `pytest tests/test_smoke.py` 
- [ ] Relevant individual smoke tests: `pytest tests/test_smoke.py::test_fill_in_the_name` 
- [ ] Backward compatibility tests: `bash tests/backward_comaptibility_tests.sh`